requirements.txt
streamlit
streamlit-option-menu


pandas
matplotlib



google-cloud-firestore